:recycle: QD-14839: unify the two download scripts into shared scripts/download-deps.go <pin>
